+++ examples/be-example-express-sqlite/CHANGELOG.md
@@ -1,5 +1,21 @@
 # be-example-express-sqlite
 
+## 0.0.18
+
+### Patch Changes
+
+- Make cache interface compatible with redis
+- Updated dependencies
+  - convostack@0.0.21
+
+## 0.0.17
+
+### Patch Changes
+
+- UI updates, backend exports
+- Updated dependencies
+  - convostack@0.0.20
+
 ## 0.0.16
 
 ### Patch Changes