+++ packages/convostack-agent-sse-echo-server/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @convostack/agent-sse-echo-server
 
+## 0.0.18
+
+### Patch Changes
+
+- Make cache interface compatible with redis
+
+## 0.0.17
+
+### Patch Changes
+
+- UI updates, backend exports
+
 ## 0.0.16
 
 ### Patch Changes