+++ packages/convostack-auth-jwt/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @convostack/auth-jwt
 
+## 0.0.18
+
+### Patch Changes
+
+- Make cache interface compatible with redis
+- Updated dependencies
+  - @convostack/auth@0.0.18
+  - @convostack/models@0.0.18
+
+## 0.0.17
+
+### Patch Changes
+
+- UI updates, backend exports
+- Updated dependencies
+  - @convostack/auth@0.0.17
+  - @convostack/models@0.0.17
+
 ## 0.0.16
 
 ### Patch Changes