+++ packages/convostack-backend-express/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @convostack/backend-express
 
+## 0.0.19
+
+### Patch Changes
+
+- Make cache interface compatible with redis
+- Updated dependencies
+  - @convostack/agent@0.0.18
+  - @convostack/auth@0.0.18
+  - @convostack/models@0.0.18
+  - @convostack/shared@0.0.18
+
+## 0.0.18
+
+### Patch Changes
+
+- UI updates, backend exports
+- Updated dependencies
+  - @convostack/agent@0.0.17
+  - @convostack/auth@0.0.17
+  - @convostack/models@0.0.17
+  - @convostack/shared@0.0.17
+
 ## 0.0.17
 
 ### Patch Changes