+++ packages/schema-graphql/CHANGELOG.md
@@ -1,5 +1,17 @@
 # schema-graphql
 
+## 0.0.16
+
+### Patch Changes
+
+- Make cache interface compatible with redis
+
+## 0.0.15
+
+### Patch Changes
+
+- UI updates, backend exports
+
 ## 0.0.14
 
 ### Patch Changes